+++ CHANGELOG.md
@@ -178,6 +178,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support artifact archive with existing `debian/` folder. During deb packages
   builds, this upstream `debian/` folder is removed and replaced by one
   generated with the artifact definition (#174).
+- Detect console unix socket closed by server, generally due to unexpected
+  `fatbuildrd` error, in order to avoid endless loop and properly stop the
+  console on client side (`fatbuildrctl` and `fatbuildrweb`) with error message.
 - docs:
   - Add missing path parameter in REST API to retrieve artifact information.
   - Add missing optional `architectures` parameter in instances pipelines